refactor(EditfeatureTypeForm): split file between multiple hooks, keep logs for now
