+++ src/features/working-layer/forms/FeatureTypeFormFields.tsx
@@ -1,0 +1,174 @@
+import { useMemo } from "react";
+import Table from "@codegouvfr/react-dsfr/Table";
+import Tooltip from "@codegouvfr/react-dsfr/Tooltip";
+import Input from "@codegouvfr/react-dsfr/Input";
+import Select from "@codegouvfr/react-dsfr/Select";
+import Checkbox from "@codegouvfr/react-dsfr/Checkbox";
+import { Upload } from "@codegouvfr/react-dsfr/Upload";
+import { FeatureTypeColumn } from "@/constants/communities/types";
+
+interface FeatureTypeFormFieldsProps {
+    columns: FeatureTypeColumn[];
+    formData: Record<string, string | number | boolean | File[] | null>;
+    validationErrors: Record<string, string | null>;
+    updateField: (name: string, value: string | number | boolean | File[] | null, col: FeatureTypeColumn) => void;
+}
+export const FeatureTypeFormFields: React.FC<FeatureTypeFormFieldsProps> = ({ columns, formData, validationErrors, updateField }) => {
+    const dataColumns = useMemo(
+        () =>
+            columns
+                .filter((col) => !col.crs)
+                .map((col) => {
+                    const v = formData[col.name];
+                    const error = validationErrors[col.name];
+
+                    const titleCell = col.description ? (
+                        <Tooltip kind="hover" title={<span>{col.description}</span>}>
+                            <span>
+                                {col.title}
+                                {col.required && <span style={{ color: "red" }}> *</span>}
+                            </span>
+                        </Tooltip>
+                    ) : (
+                        <span>
+                            {col.title}
+                            {col.required && <span style={{ color: "red" }}> *</span>}
+                        </span>
+                    );
+
+                    let valueCell: React.ReactNode;
+
+                    if (col.read_only) {
+                        valueCell = (
+                            <Input
+                                label=""
+                                nativeInputProps={{
+                                    value: String(v ?? ""),
+                                    disabled: true,
+                                }}
+                            />
+                        );
+                    } else {
+                        switch (col.type.toLowerCase()) {
+                            case "string":
+                                valueCell = (
+                                    <Input
+                                        label=""
+                                        state={error ? "error" : "default"}
+                                        stateRelatedMessage={error || undefined}
+                                        nativeInputProps={{
+                                            value: (v as string) ?? "",
+                                            onChange: (e) => updateField(col.name, e.target.value, col),
+                                            required: col.required,
+                                            minLength: col.min_length ?? undefined,
+                                            maxLength: col.max_length ?? undefined,
+                                            pattern: col.pattern ?? undefined,
+                                        }}
+                                    />
+                                );
+                                break;
+
+                            case "integer":
+                                valueCell = col.enum ? (
+                                    <Select
+                                        label=""
+                                        state={error ? "error" : "default"}
+                                        stateRelatedMessage={error || undefined}
+                                        nativeSelectProps={{
+                                            value: (v as string) ?? "",
+                                            onChange: (e) => updateField(col.name, e.target.value, col),
+                                            required: col.required,
+                                        }}
+                                    >
+                                        <option value="">Sélectionnez une option</option>
+                                        {col.enum.map((opt, idx) => (
+                                            <option key={idx} value={opt}>
+                                                {opt}
+                                            </option>
+                                        ))}
+                                    </Select>
+                                ) : (
+                                    <Input
+                                        label=""
+                                        state={error ? "error" : "default"}
+                                        stateRelatedMessage={error || undefined}
+                                        nativeInputProps={{
+                                            type: "number",
+                                            value: (v as number | string) ?? "",
+                                            onChange: (e) => updateField(col.name, Number(e.target.value), col),
+                                            required: col.required,
+                                            min: col.min_value ?? undefined,
+                                            max: col.max_value ?? undefined,
+                                        }}
+                                    />
+                                );
+                                break;
+
+                            case "boolean":
+                                valueCell = (
+                                    <Checkbox
+                                        state={error ? "error" : "default"}
+                                        stateRelatedMessage={error || undefined}
+                                        options={[
+                                            {
+                                                label: "",
+                                                nativeInputProps: {
+                                                    checked: Boolean(v),
+                                                    onChange: (e) => updateField(col.name, e.target.checked, col),
+                                                    required: col.required,
+                                                },
+                                            },
+                                        ]}
+                                    />
+                                );
+                                break;
+
+                            case "date":
+                                valueCell = (
+                                    <Input
+                                        label=""
+                                        state={error ? "error" : "default"}
+                                        stateRelatedMessage={error || undefined}
+                                        nativeInputProps={{
+                                            type: "date",
+                                            value: (v as string) ?? "",
+                                            onChange: (e) => updateField(col.name, e.target.value, col),
+                                            required: col.required,
+                                        }}
+                                    />
+                                );
+                                break;
+
+                            case "document":
+                                valueCell = (
+                                    <Upload
+                                        label=""
+                                        multiple
+                                        className="upload-file"
+                                        state={error ? "error" : "default"}
+                                        stateRelatedMessage={error || undefined}
+                                        nativeInputProps={{
+                                            onChange: (e) => updateField(col.name, Array.from(e.target.files ?? []), col),
+                                            required: col.required,
+                                        }}
+                                    />
+                                );
+                                break;
+
+                            default:
+                                valueCell = <></>;
+                        }
+                    }
+
+                    return [
+                        titleCell,
+                        <div className="feature-type-form-input-cell" key={col.name}>
+                            {valueCell}
+                        </div>,
+                    ];
+                }),
+        [columns, formData, validationErrors, updateField]
+    );
+
+    return <Table bordered fixed className="feature-type-form-table" data={dataColumns} />;
+};
